third_party/blink/web_tests/platform/generic/svg/animations/animate-mpath-insert-expected.txt

SVG 1.1 dynamic animation tests

test to determine whether inserting mpath dynamically works

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS rootSVGElement.getBBox().x is 100
PASS rootSVGElement.getBBox().y is 250
PASS rootSVGElement.getBBox().x is 400
PASS rootSVGElement.getBBox().y is 250
PASS successfullyParsed is true

TEST COMPLETE
